Fix issue with date picker picking a year at the end of the range and the date being rejected because the calendar day took the selection outside of range
When the date picker is in use, it is one frequent scenario to leave the date
in the calendar as it is, go to change the year, and when that is set to come
back and adjust the date.  If the date range is limited with minDate and/or
maxDate, it is quite possible that selecting the year will mean that leaving the
calendar selected date unchanged will take you outside of the range of allowable
dates.

The current solution is to return you to the old date.  This is confusing, and
when you find out what has happened frustrating for the user.  So what this
commit does is when the date change takes you outside of the allowable range, it
to change the date to the minDate, if the selection was before the minDate and
to change it to maxDate if the selection was after the maxDate.

This is far less confusing to the user, but still maintains the integrity of the
dates within the permissable range.